requirements.txt
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
clickclick==1.2.2
connexion==2.7.0
Flask==1.1.2
idna==2.9
importlib-metadata==1.6.0
inflection==0.4.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
MarkupSafe==1.1.1
more-itertools==8.2.0
numpy==1.18.3
openapi-spec-validator==0.2.8
packaging==20.3
pluggy==0.13.1
py==1.8.1
pymongo==3.10.1
pyparsing==2.4.7
pyrsistent==0.16.0
pytest==5.4.1
pytest-asyncio==0.11.0
python-dotenv==0.13.0
PyYAML==5.3.1
requests==2.23.0
six==1.14.0
swagger-ui-bundle==0.0.6
urllib3==1.25.9
wcwidth==0.1.9
Werkzeug==1.0.1
zipp==3.1.0
